+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.0.9]
 - [**Bug**] don't set default python3 path (if user adds it, he has to install dothttp-req)
+- [**Bug**] postman import is creating duplicate folder, use showOpenDialog rather than, showSaveDialog
 - [**Bug**] history execute is not showing file extension/ file syntax
 - [**Bug**] history item hove is showing `200 undefined` fixed.
 - [**Improvement**] history item will now start showing time it executed for better difference